Program that prints various types

+++ 0x00-hello_world/6-size.c
@@ -11,10 +11,10 @@ int main(void)
 	long long int d;
 	float e;
 
-	printf("Size of char: %d bytes(s)\n", sizeof(a));
-	printf("Size of int: %d bytes(s)\n", sizeof(b));
-	printf("Size of long int: %d bytes(s)\n", sizeof(c));
-	printf("Size of long long int: %d bytes(s)\n", sizeof(d));
-	printf("Size of float: %d bytes(s)\n", sizeof(e));
+	printf("Size of char: %lu bytes(s)\n", (unsigned long)sizeof(a));
+	printf("Size of int: %lu bytes(s)\n", (unsigned long)sizeof(b));
+	printf("Size of long int: %lu bytes(s)\n", (unsigned long)sizeof(c));
+	printf("Size of long long int: %lu bytes(s)\n", (unsigned long)sizeof(d));
+	printf("Size of float: %lu bytes(s)\n", (unsigned long)sizeof(e));
 	return (0);
 }
